Add ability to set ZFS properties (#41)
This commit adds a generic tool to convert various python
objects into an nvlist of ZFS properties and plumbs it into
the set_properties method for ZFSResource objects. The
common nvlist code will also be used in other places that
allow bulk-setting of ZFS properties (for example when
creating new datasets).