NAS-136891 / 25.10 / Add method to change crypto key (#76)
This commit adds a change_key method to the ZFS cryptography object,
which may be used to rewrap the crypto key on an existing encrypted ZFS
resource (dataset or volume). It also creates a new struct sequence
object for cryptography information that will be consumed when creating
new encrypted resources.