Add retry logic for zfs_userspace iterator
If a dataset is unmounted then the zfs_userspace related ZFS
ioctls may fail with EZFS_BUSY depending on timing. In this
case we can generally retry after a slight wait and succeed.
This commit adds logic to usleep for up to 5 seconds while
trying to start iteration. This addresses potential spurious
errors that can filter up to library consumers such as the
TrueNAS middleware / UI if multiple concurrent requests are
issued for different quota types on the same unmounted dataset.