Add zpool status method for pool objects
This commit adds the a new method to generate status of a zpool.
The status() method generates a struct sequence object containing
various health-related details for the pools and its vdevs.

As part of this commit, a new module `truenas_pylibzfs.enums` was
added and references to our various ZFS enums were placed in it
with the view of eventually eliminating them from root module.