NAS-137087 / 25.10-RC.1 / Add channel program wrapper (by anodos325) (#99)
This commit adds ability to execute ZFS channel programs with GIL
dropped. The channel program contents are audited through the python
audit module (if library consumer is using it) and a zpool history entry
is generated if write access to pool was requested for the channel
program.

Original PR: https://github.com/truenas/truenas_pylibzfs/pull/89