NAS-137285 / 25.10-RC.1 / Allow deletion of multiple snapshots in same ds (by anodos325) (#107)
This commit fixes validation that was preventing concurrent snapshot
deletion of multiple snapshots in the same dataset. The specific way the
issue is addressed is by changing the callback function used when
iterating sets of holds / snapshots so that it takes an operation type
enum argument and using it to determine whether to apply the extra
validation required for creating new snapshots.

Original PR: https://github.com/truenas/truenas_pylibzfs/pull/106

Co-authored-by: Andrew Walker <andrew.walker@truenas.com>