NAS-137818 / 26.04 / Add more zfs_prop_get error handling (#112)
If dds_inconsistent due to in-progress replication some ZFS properties
may not be retrievable. In this case we should set the property value to
something indicating that there is a transient issue, but not error out
entirely. For this purpose we'll use the parsed value of None type, with
a raw value of string `"<INCONSISTENT>"`.