NAS-137818 / 25.10.0 / Add more zfs_prop_get error handling (by anodos325) (#113)
If dds_inconsistent due to in-progress replication some ZFS properties
may not be retrievable. In this case we should set the property value to
something indicating that there is a transient issue, but not error out
entirely. For this purpose we'll use the parsed value of None type, with
a raw value of string `"<INCONSISTENT>"`.

Original PR: https://github.com/truenas/truenas_pylibzfs/pull/112

---------

Co-authored-by: Andrew Walker <andrew.walker@truenas.com>