Add zpool event iterator
This commit adds the ability to open a zpool events iterator
through the zpool_events() method on an open handle.

For example:

```
import truenas_pylibzfs
z = truenas_pylibzfs.open_handle()
for i in z.zpool_events():
    print(i)
```

Currently we support seeking to the end of the events and then
iterating from that point, and setting a blocking mode where
we wait with GIL not held until events occur.